put submit_work_detail into a helper to avoid the duplicate codes.
